Add focus section
This commit adds the my work in focus section, with Strifecake as the first project. This commit also adds scaffolding for other projects which will be added in future commits. This part of the README relies heavily on collapsible sections to ensure it will not take up too much vertical space. Since it currently only holds one section, it is expanded by default. In the future, only the most relevant showcase project will be expanded whereas the other projects will be collapsed.

* Create scaffolding for showcase projects
* Add Strifecake anchor CMS theme showcase